refactor-compact-tiles * Create component out of SubLabelIcon from inside TileRow. Use same SubLabelIcon in both compact/TileRow and ChronoTableRows
