+++ src/components/SubLabelIcon/SubLabelIcon.module.scss
@@ -1,0 +1,23 @@
+
+.Cancellation {
+    align-items: center;
+    border-radius: 0.5rem;
+    display: flex;
+    height: 0.875rem;
+    justify-content: center;
+    margin-left: 0.375rem;
+    overflow: visible;
+    width: 0.875rem;
+}
+
+
+.Situation {
+    font-size: 1.125rem;
+    height: 1.25rem;
+    margin-left: 0.375rem;
+}
+
+.Icon {
+    margin-left: -0.125rem;
+    margin-right: -0.125rem;
+}
