+++ src/dashboards/Chrono/ChronoTileRows/ChronoTableRows.tsx
@@ -1,16 +1,12 @@
 import React, { Fragment } from 'react'
 import { Heading3 } from '@entur/typography'
 import { DataCell, TableBody, TableRow } from '@entur/table'
-import { ValidationExclamation } from '../../../assets/icons/ValidationExclamation'
-import { ValidationError } from '../../../assets/icons/ValidationError'
-import { IconColorType, LineData, TileSubLabel } from '../../../types'
-import { SituationModal } from '../../../components/SituationModal/SituationModal'
-import { createTileSubLabel, isMobileWeb } from '../../../utils/utils'
+import { IconColorType, LineData } from '../../../types'
+import { createTileSubLabel } from '../../../utils/utils'
 import { getIcon } from '../../../utils/icon'
 import { NewDayTableRow } from '../../../components/DateRow/NewDayTableRow'
+import { SubLabelIcon } from '../../../components/SubLabelIcon/SubLabelIcon'
 import classes from './ChronoTableRows.module.scss'
-
-const isMobile = isMobileWeb()
 
 interface ChronoTableRowsProps {
     visibleDepartures: LineData[]
@@ -75,36 +71,6 @@ function ChronoTableRows({
             })}
         </TableBody>
     )
-}
-
-function SubLabelIcon({
-    subLabel,
-    hideSituations,
-}: {
-    subLabel: TileSubLabel
-    hideSituations?: boolean
-}): JSX.Element | null {
-    if (!hideSituations && subLabel?.situation)
-        if (isMobile)
-            return (
-                <div>
-                    <SituationModal situationMessage={subLabel.situation} />
-                </div>
-            )
-        else
-            return (
-                <div>
-                    <ValidationExclamation />
-                </div>
-            )
-
-    if (subLabel.hasCancellation)
-        return (
-            <div className={classes.Cancellation}>
-                <ValidationError />
-            </div>
-        )
-    return null
 }
 
 export { ChronoTableRows }
